+++ app/views/spina/admin/shared/_navigation.html.erb
@@ -41,7 +41,7 @@
           <%= render Spina::MainNavigation::LinkComponent.new("General", spina.edit_admin_account_path, active: controller_name == "accounts") %>
           <%= render Spina::MainNavigation::LinkComponent.new(t('spina.website.theme'), spina.edit_admin_theme_path, active: controller_name == "theme") %>
           
-          <% if Spina.config.authentication == "Spina::Authentication" %>
+          <% if Spina.config.authentication == "Spina::Authentication::Sessions" %>
             <%= render Spina::MainNavigation::LinkComponent.new("Users", spina.admin_users_path, active: controller_name == "users") %>
           <% end %>
