+++ lib/spina.rb
@@ -7,8 +7,8 @@ require 'spina/theme'
 require 'spina/tailwind_purger'
 require 'spina/attr_json_spina_parts_model'
 require 'spina/attr_json_monkeypatch'
-require 'spina/authentication'
-require 'spina/basic_authentication'
+require 'spina/authentication/sessions'
+require 'spina/authentication/basic'
 
 module Spina
   include ActiveSupport::Configurable
@@ -33,8 +33,8 @@ module Spina
   
   # Specify the module that handles authentication
   # You can swap this out for something like Devise, or you can use your own authentication.
-  # The default is Spina::Authentication and includes basic user management
-  self.authentication = "Spina::Authentication"
+  # The default is Spina::Authentication::Sessions and includes basic user management
+  self.authentication = "Spina::Authentication::Sessions"
   
   # The parent controller all frontend Spina controllers inherit from
   # Default is ApplicationController
