+++ lib/spina/authentication/basic.rb
@@ -1,0 +1,24 @@
+module Spina
+  module Authentication
+    module Basic
+      extend ActiveSupport::Concern
+      
+      included do
+        helper_method :logged_in?
+      end
+      
+      def logged_in?
+        authenticate
+      end
+      
+      private
+      
+        def authenticate
+          authenticate_or_request_with_http_basic do |username, password|
+            "spina" == username && "demo" == password
+          end
+        end
+    
+    end
+  end
+end
